requirements.txt
streamlit
python-dotenv
langchain-core
langchain-groq
langgraph
graphviz
fpdf
pydantic
typing-extensions